(feature) GitLabStatus understands isMaster()
Instead of asserting on the type of the object when deciding if a model
is a masterStatus, ask it.

This method is not tested because it is difficult to test parse().
When sinon tests are added this will become simpler.

Also removes isThumbsUp ranking, which should have been taken out at the
beginning of this branch.