(feature) Allow fetching repo list from bare JSON
Repo lists heretofore had to be stored in GitHub and the Fourth Wall
method for reading from these files depended on the structure of the
response from the GitHub server. This feature lets a user send a json=
param into the query string and have that file parsed straight into a
repo list.